+++ packages/app/src/cli/services/function/ui/components/hooks/setup-extension-watcher-for-replay.ts
